frontend/djangobackend/requirements.txt
asgiref==3.2.7
dataclasses==0.8
Django==3.0.5
django-cors-headers==3.3.0
djangorestframework==3.11.0
djongo==1.3.3
pkg-resources==0.0.0
pymongo==3.11.1
pytz==2020.1
sqlparse==0.2.4
